+++ extensions/browser-bookmarks/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Browser Bookmarks Changelog
 
+## [Support for Vivaldi Snapshot Browser] - 2026-02-08
+
+- Added support for `Vivaldi Snapshot` browser.
+
 ## [Support for Helium Browser] - 2025-12-08
 
 - Added support for `Helium` browser.